change 'component push' to 'push'
Push is now a root level command as specified in use cases.
fixes #86
